# src/Happy-Spirit-Games_StoryTimeAPI_2.0.0_swagger.yaml
openapi: 3.0.1
info:
  title: StoryTime Service API
  description: >-
    Application programming interface (API) to the StoryTime service, supporting clients and consumers of interactive fiction stories where the player decides what path to take through the story.
  termsOfService: https://happyspiritgames.blog/terms-of-service/
  contact:
    name: Happy Spirit Games Support
    url: https://happyspiritgames.blog/terms-of-service/
    email: welcome@happyspiritgames.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "2.0.0"
servers:
  - url: https://virtserver.swaggerhub.com/Happy-Spirit-Games/StoryTimeReaderAPI/2.0.0
    description: API definition design and testing
  - url: https://hsg-storytime.herokuapp.com/api
    description: Production StoryTime service
  - url: https://hsg-storytime-staging.herokuapp.com/api
    description: Staging StoryTime service for pre-production testing
paths:
...
  /draft-stories:
    get:
      tags:
        - author
      summary: Returns all draft stories owned by the current author.
      description: Returns a list of stories in draft and owned by the user who is logged in. That means they are open for changes.
      operationId: getDraftSummaries
      responses:
        '200':
          description: Draft stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DraftStorySummary'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - author
      summary: Creates new draft story
      description: Creates a story in draft mode that is owned by the logged in user, a.k.a. the author.  Being in draft means that the story is open for editing.  Note that firstSceneId is assigned by the system.  Any value supplied by the client will be ignored.
      operationId: beginNewStory
      requestBody:
        description: Story summary information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftStorySummaryUpdate'
      responses:
        '201':
          description: Created draft story
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftStorySummary'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /draft-stories/{storyId}:
    get:
      tags:
        - author
      summary: Returns draft stories with the given ID.
      description: Returns draft stories with the given ID.  The story must be owned by the user.
      operationId: getStorySummary
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Draft story
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftStorySummary'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - author
      summary: Updates a draft story
      description: Updates a story that is in draft mode and owned by the logged in user, a.k.a. the author.
      operationId: updateStorySummary
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
      requestBody:
        description: Story summary information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftStorySummaryUpdate'
      responses:
        '202':
          description: Accepted changes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftStorySummary'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /draft-stories/{storyId}/full:
    get:
      tags:
        - author
      summary: Returns draft stories with the given ID.
      description: Returns draft stories with the given ID.  The story must be owned by the user.
      operationId: getStorySummary
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Draft story
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullDraftStory'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /draft-stories/{storyId}/publish:
    post:
      tags:
        - author
      summary: Creates a new published version of the story.
      description: Creates a new published version of the story.  The publishing process is TBD.  There will most likely be a review and approval step.  Maybe that can come later.  Regardless, this method makes a copy of the story that is ready to play in the StoryTime Reader.
      operationId: startPublishingProcess
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Created published story
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /draft-stories/{storyId}/scenes:
    post:
      tags:
        - author
      summary: Creates new draft scene
      description: Creates a scene in draft mode for the given story.  Being in draft means that the scene is open for editing.
      operationId: beginNewScene
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
      requestBody:
        description: Scene information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftSceneUpdate'
      responses:
        '201':
          description: Created scene
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftScene'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /draft-stories/{storyId}/scenes/{sceneId}:
    get:
      tags:
        - author
      summary: Updates a scene in draft mode.
      description: Updates a scene in draft mode.
      operationId: updateScene
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
        - name: sceneId
          in: path
          description: scene ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Found scene
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftScene'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - author
      summary: Updates a scene in draft mode.
      description: Updates a scene in draft mode.
      operationId: updateScene
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
        - name: sceneId
          in: path
          description: scene ID
          required: true
          schema:
            type: string
      requestBody:
        description: Scene information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftSceneUpdate'
      responses:
        '202':
          description: Accepted changes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftScene'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /draft-stories/{storyId}/scenes/{sceneId}/signpost:
    get:
      tags:
        - author
      summary: Gets the signpost for a draft scene.
      description: Gets the signpost for a draft scene.  The signpost lists the options for where the player can go next.
      operationId: getSignpost
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
        - name: sceneId
          in: path
          description: scene ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Found scene
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftSignpost'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - author
      summary: Updates a signpost on a scene for a story in draft mode.
      description: Updates a signpost on a scene for a story in draft mode.  This method handles all of the ways to modify a signpost.  Including a new destination in toUpdate will create a new sign.  Including an existing destination will update the sign.  Adding a destination to the toDelete array will remove the sign.  Removal happens first, but there is no need to remove a destination that is being updated.  It will simply be recreated.  If you are trying to remove a destination, be sure not to update it in the same call.
      operationId: updateSignpost
      parameters:
        - name: storyId
          in: path
          description: story ID
          required: true
          schema:
            type: string
        - name: sceneId
          in: path
          description: scene ID
          required: true
          schema:
            type: string
      requestBody:
        description: Scene information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftSignpostUpdate'
      responses:
        '202':
          description: Accepted changes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftSignpost'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    DraftScene:
      type: object
      properties:
        sceneId:
          type: string
        title:
          type: string
        prose:
          type: string
        endPrompt:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        signpost:
          $ref: '#/components/schemas/DraftSignpost'
    DraftSceneUpdate:
      type: object
      properties:
        title:
          type: string
        prose:
          type: string
        endPrompt:
          type: string
    DraftSign:
      type: object
      properties:
        destinationId:
          type: string
        teaser:
          type: string
        order:
          type: integer
    DraftSignpost:
      type: array
      items:
        $ref: '#/components/schemas/DraftSign'
    DraftSignpostUpdate:
      type: object
      properties:
        toUpdate:
          $ref: '#/components/schemas/DraftSignpost'
        toDelete:
          type: array
          items:
            type: string
    DraftStorySummary:
      type: object
      properties:
        storyId:
          type: string
        authorId:
          type: string
          format: uuid
        title:
          type: string
        tagLine:
          type: string
        about:
          type: string
        firstSceneId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DraftStorySummaryUpdate:
      type: object
      properties:
        title:
          type: string
        tagLine:
          type: string
        about:
          type: string
        firstSceneId:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    FullDraftStory:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/DraftStorySummary'
        scenes:
          type: array
          items:
            $ref: '#/components/schemas/DraftScene'
...